+++ src/main/java/com/app/controller/UserController.java
@@ -46,7 +46,7 @@ public class UserController {
 
 	}
 
-	public ResponseDTO<?> updateProfile(@PathVariable Long userid, @RequestBody UpdateDTO user) {
+	public ResponseDTO<?> updateProfile(@PathVariable Integer userid, @RequestBody UpdateDTO user) {
 		try {
 			return new ResponseDTO<>(HttpStatus.OK, "Update operation done successfully",
 					userService.updateProfile(userid, user));
@@ -61,13 +61,13 @@ public class UserController {
 	}
 
 	@GetMapping("/facility/{propid}")
-	public ResponseDTO<?> getAllPropertyHomePage(@PathVariable Long propid) {
+	public ResponseDTO<?> getAllPropertyHomePage(@PathVariable Integer propid) {
 		return new ResponseDTO<>(HttpStatus.OK, "Fetching property list successfully",
 				userService.getPropertyFacilities(propid));
 	}
 
 	@PostMapping("/profile/{userid}")
-	public ResponseDTO<?> profilePage(@PathVariable Long userid, @RequestBody Role role) {
+	public ResponseDTO<?> profilePage(@PathVariable Integer userid, @RequestBody Role role) {
 		return new ResponseDTO<>(HttpStatus.OK, "Fetching property list successfully",
 				userService.profilePage(userid, role));
 	}
