+++ src/main/java/com/app/dao/FacilityRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.pojos.Facilities;
 
-public interface FacilityRepository extends JpaRepository<Facilities,Long> {
+public interface FacilityRepository extends JpaRepository<Facilities,Integer> {
 
 }
