+++ src/main/java/com/app/services/ISeekerService.java
@@ -5,10 +5,10 @@ import java.util.List;
 import com.app.custom_excs.ResourceNotFoundException;
 import com.app.dto.OwnerDetailsDTO;
 import com.app.dto.PropertyDetailsDTO;
+//import com.app.dto.AreaDTO;
 
 public interface ISeekerService {
-	OwnerDetailsDTO getContactDetails(Long propid) throws ResourceNotFoundException;
-
-	List<PropertyDetailsDTO> searchPropertyByCity(String city) throws ResourceNotFoundException;
+           OwnerDetailsDTO getContactDetails(int propid) throws ResourceNotFoundException;
+       List<PropertyDetailsDTO> searchPropertyByCity(String city) throws ResourceNotFoundException;
 
 }
