+++ src/main/java/com/app/services/IUserService.java
@@ -13,10 +13,10 @@ import com.app.pojos.User;
 public interface IUserService {
 	User loginRequest(String email,String password);
 	User registerationForm(User user);
-	User updateProfile(Long userid,UpdateDTO user) throws ResourceNotFoundException;
+	User updateProfile(int userid,UpdateDTO user) throws ResourceNotFoundException;
 	List<PropertyDetailsDTO> getAllProperty();
-	FacilityDetailsDTO getPropertyFacilities(Long propid);
-	User profilePage(Long userid,Role role);
+	FacilityDetailsDTO getPropertyFacilities(int propid);
+	User profilePage(int userid,Role role);
 	  String forgetPassword(String email,String password) throws ResourceNotFoundException;
