+++ src/main/java/com/app/services/SeekerServiceImpl.java
@@ -1,7 +1,6 @@
 package com.app.services;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,7 +12,6 @@ import com.app.dao.PropertyRepository;
 import com.app.dao.UserRepository;
 import com.app.dto.OwnerDetailsDTO;
 import com.app.dto.PropertyDetailsDTO;
-
 //import com.app.dto.AreaDTO;
 @Service
 @Transactional
@@ -22,24 +20,22 @@ public class SeekerServiceImpl implements ISeekerService {
 	private UserRepository userRepo;
 	@Autowired
 	private PropertyRepository propRepo;
-
 	@Override
-	public OwnerDetailsDTO getContactDetails(Long propid) throws ResourceNotFoundException {
-		boolean exists = propRepo.existsById(propid);
-		if (!exists)
+	public OwnerDetailsDTO getContactDetails(int propid) throws ResourceNotFoundException {
+		boolean exists=propRepo.existsById(propid);
+		   if(!exists)
 			throw new ResourceNotFoundException("property is unavailable !!!");
-		OwnerDetailsDTO contact = propRepo.getOwnerDetails(propid);
-		return contact;
-
-	}
-
+		     OwnerDetailsDTO contact=propRepo.getOwnerDetails(propid);
+		       return contact;
+              
+}
 	@Override
-	public List<PropertyDetailsDTO> searchPropertyByCity(String city) throws ResourceNotFoundException {
-		String name = propRepo.existsCityByName(city);
-		boolean exists = Boolean.parseBoolean(name);
-		if (exists)
-			throw new ResourceNotFoundException("Invalid cityName and areaName!!!!!");
-		return propRepo.findFlatByCity(city);
-
+	public List<PropertyDetailsDTO> searchPropertyByCity( String city) throws  ResourceNotFoundException{
+		         String name=propRepo.existsCityByName(city);
+		         boolean exists=Boolean.parseBoolean(name);
+		         if(exists)
+		         throw new ResourceNotFoundException("Invalid cityName and areaName!!!!!");
+		         return propRepo.findFlatByCity( city);
+		       
 	}
 }
